+++ src/lib.rs
@@ -8,6 +8,101 @@
 //! For more information see [https://nats.io/].
 //!
 //! [https://nats.io/]: https://nats.io/
+//!
+//! ## Examples
+//!
+//! `> cargo run --example nats-box -- -h`
+//!
+//! Basic connections, and those with options. The compiler will force these to be correct.
+//!
+//! ```no_run
+//! # fn main() -> Result<(), Box<dyn std::error::Error>> {
+//! let nc = nats::connect("demo.nats.io")?;
+//!
+//! let nc2 = nats::ConnectionOptions::new()
+//!     .with_name("My Rust NATS App")
+//!     .with_user_pass("derek", "s3cr3t!")
+//!     .connect("127.0.0.1")?;
+//!
+//! let nc3 = nats::ConnectionOptions::new()
+//!     .with_credentials("path/to/my.creds")
+//!     .connect("connect.ngs.global")?;
+//!
+//! let tls_connector = nats::tls::builder()
+//!     .identity(nats::tls::Identity::from_pkcs12(b"der_bytes", "my_password")?)
+//!     .add_root_certificate(nats::tls::Certificate::from_pem(b"my_pem_bytes")?)
+//!     .build()?;
+//!
+//! let nc4 = nats::ConnectionOptions::new()
+//!     .tls_connector(tls_connector)
+//!     .connect("tls://demo.nats.io:4443")?;
+//! # Ok(()) }
+//! ```
+//!
+//! ### Publish
+//!
+//! ```no_run
+//! # fn main() -> std::io::Result<()> {
+//! let nc = nats::connect("demo.nats.io")?;
+//! nc.publish("my.subject", "Hello World!")?;
+//!
+//! nc.publish("my.subject", "my message")?;
+//!
+//! // Publish a request manually.
+//! let reply = nc.new_inbox();
+//! let rsub = nc.subscribe(&reply)?;
+//! nc.publish_request("my.subject", &reply, "Help me!")?;
+//! # Ok(()) }
+//! ```
+//!
+//! ### Subscribe
+//!
+//! ```no_run
+//! # fn main() -> std::io::Result<()> {
+//! # use std::time::Duration;
+//! let nc = nats::connect("demo.nats.io")?;
+//! let sub = nc.subscribe("foo")?;
+//! for msg in sub.messages() {}
+//!
+//! // Using next.
+//! if let Some(msg) = sub.next() {}
+//!
+//! // Other iterators.
+//! for msg in sub.try_iter() {}
+//! for msg in sub.timeout_iter(Duration::from_secs(10)) {}
+//!
+//! // Using a threaded handler.
+//! let sub = nc.subscribe("bar")?.with_handler(move |msg| {
+//!     println!("Received {}", &msg);
+//!     Ok(())
+//! });
+//!
+//! // Queue subscription.
+//! let qsub = nc.queue_subscribe("foo", "my_group")?;
+//! # Ok(()) }
+//! ```
+//!
+//! ### Request/Response
+//!
+//! ```no_run
+//! # use std::time::Duration;
+//! # fn main() -> std::io::Result<()> {
+//! let nc = nats::connect("demo.nats.io")?;
+//! let resp = nc.request("foo", "Help me?")?;
+//!
+//! // With a timeout.
+//! let resp = nc.request_timeout("foo", "Help me?", Duration::from_secs(2))?;
+//!
+//! // With multiple responses.
+//! for msg in nc.request_multi("foo", "Help")?.iter() {}
+//!
+//! // Publish a request manually.
+//! let reply = nc.new_inbox();
+//! let rsub = nc.subscribe(&reply)?;
+//! nc.publish_request("foo", &reply, "Help me!")?;
+//! let response = rsub.iter().take(1);
+//! # Ok(()) }
+//! ```
 
 #![recursion_limit = "1024"]
 // #![cfg_attr(test, deny(warnings))]
@@ -67,7 +162,6 @@
     clippy::print_stdout,
     clippy::pub_enum_variant_names,
     clippy::redundant_closure_for_method_calls,
-    clippy::replace_consts,
     clippy::result_map_unwrap_or_else,
     clippy::shadow_reuse,
     clippy::shadow_same,
@@ -84,6 +178,8 @@
     clippy::wildcard_enum_match_arm,
     clippy::wrong_pub_self_convention,
 )]
+
+const DEFAULT_FLUSH_TIMEOUT: Duration = Duration::from_secs(10);
 
 mod connect;
 mod creds_utils;
@@ -646,7 +742,7 @@ impl<TypeState> ConnectionOptions<TypeState> {
     /// # Examples
     /// ```no_run
     /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// let mut tls_connector = nats::tls::builder()
+    /// let tls_connector = nats::tls::builder()
     ///     .identity(nats::tls::Identity::from_pkcs12(b"der_bytes", "my_password")?)
     ///     .add_root_certificate(nats::tls::Certificate::from_pem(b"my_pem_bytes")?)
     ///     .build()?;
@@ -857,7 +953,7 @@ impl Connection {
         let sid = self.sid.fetch_add(1, Ordering::Relaxed);
         self.shared_state
             .outbound
-            .send_sub_msg(subject, queue, sid)?;
+            .send_sub_msg(subject, queue, sid, &self.shared_state)?;
         let (sub, recv) = crossbeam_channel::unbounded();
         {
             let mut subs = self.shared_state.subs.write();
@@ -894,7 +990,7 @@ impl Connection {
     pub fn publish(&self, subject: &str, msg: impl AsRef<[u8]>) -> io::Result<()> {
         self.shared_state
             .outbound
-            .send_pub_msg(subject, None, msg.as_ref())
+            .send_pub_msg(subject, None, msg.as_ref(), &self.shared_state)
     }
 
     /// Publish a message on the given subject with a reply subject for responses.
@@ -915,9 +1011,12 @@ impl Connection {
         reply: &str,
         msg: impl AsRef<[u8]>,
     ) -> io::Result<()> {
-        self.shared_state
-            .outbound
-            .send_pub_msg(subject, Some(reply), msg.as_ref())
+        self.shared_state.outbound.send_pub_msg(
+            subject,
+            Some(reply),
+            msg.as_ref(),
+            &self.shared_state,
+        )
     }
 
     /// Create a new globally unique inbox which can be used for replies.
@@ -1015,7 +1114,7 @@ impl Connection {
     /// # }
     /// ```
     pub fn flush(&self) -> io::Result<()> {
-        self.flush_timeout(Duration::from_secs(10))
+        self.flush_timeout(DEFAULT_FLUSH_TIMEOUT)
     }
 
     /// Flush a NATS connection by sending a `PING` protocol and waiting for the responding `PONG`.
@@ -1032,33 +1131,7 @@ impl Connection {
     /// # }
     /// ```
     pub fn flush_timeout(&self, duration: Duration) -> io::Result<()> {
-        let (s, r) = crossbeam_channel::bounded(1);
-
-        // We take out the mutex before sending a ping (which may fail)
-        inject_delay();
-        let mut pongs = self.shared_state.pongs.lock();
-
-        // This will throw an error if the system is disconnected.
-        self.shared_state.outbound.send_ping()?;
-
-        // We only push to the mutex if the ping was successfully queued.
-        // By holding the mutex across calls, we guarantee ordering in the
-        // queue will match the order of calls to `send_ping`.
-        pongs.push_back(s);
-        drop(pongs);
-
-        let success = r
-            .recv_timeout(duration)
-            .map_err(|_| Error::new(ErrorKind::TimedOut, "No response"))?;
-
-        if !success {
-            return Err(Error::new(
-                ErrorKind::BrokenPipe,
-                "The connection to the remote server was lost",
-            ));
-        }
-
-        Ok(())
+        self.shared_state.flush_timeout(duration)
     }
 
     /// Close a NATS connection. All clones of
@@ -1096,7 +1169,7 @@ impl Connection {
     /// ```
     pub fn rtt(&self) -> io::Result<Duration> {
         let start = Instant::now();
-        self.flush_timeout(Duration::from_secs(10))?;
+        self.flush_timeout(DEFAULT_FLUSH_TIMEOUT)?;
         Ok(start.elapsed())
     }
 
@@ -1150,5 +1223,42 @@ impl Connection {
     pub fn client_id(&self) -> u64 {
         let info = self.shared_state.info.read();
         info.client_id
+    }
+
+    /// Send an unsubscription for all subs then flush the connection, allowing any unprocessed
+    /// messages to be handled by a handler function if one is configured.
+    ///
+    /// After the flush returns, we know that a round-trip to the server has happened after it
+    /// received our unsubscription, so we shut down the subscriber afterwards.
+    ///
+    /// A similar method exists for the `Subscription` struct which will drain
+    /// a single `Subscription` without shutting down the entire connection
+    /// afterward.
+    ///
+    /// # Example
+    /// ```
+    /// # use std::sync::{Arc, atomic::{AtomicBool, Ordering::SeqCst}};
+    /// # fn main() -> std::io::Result<()> {
+    /// # let nc = nats::connect("demo.nats.io")?;
+    /// let received = Arc::new(AtomicBool::new(false));
+    /// let received_2 = received.clone();
+    ///
+    /// nc.subscribe("test.drain")?.with_handler(move |m| {
+    ///     received_2.store(true, SeqCst);
+    ///     Ok(())
+    /// });
+    ///
+    /// nc.publish("test.drain", "message")?;
+    /// nc.drain()?;
+    ///
+    /// # std::thread::sleep(std::time::Duration::from_secs(1));
+    ///
+    /// assert!(received.load(SeqCst));
+    ///
+    /// # Ok(())
+    /// # }
+    /// ```
+    pub fn drain(&self) -> io::Result<()> {
+        self.shared_state.drain()
     }
 }