+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# Unreleased
+# 0.4.0
+
+## New Features
+
+* #57 The `drain` method has been added to the
+  `Connection` and `Subscription` structs.
 
 ## Breaking Changes